Make DrawTriangle to accept fp coordinates.

+++ gfx/triangle.h
@@ -4,6 +4,6 @@
 #include "gfx/canvas.h"
 
 void DrawTriangle(CanvasT *canvas,
-                  int x1, int y1, int x2, int y2, int x3, int y3);
+                  float x1, float y1, float x2, float y2, float x3, float y3);
 
 #endif
